HT_23/requirements.txt
Django==4.1.7
requests==2.28.2
djangorestframework==3.14.0
celery==5.2.7
redis==4.5.1
